refactor(core): move state and timestamp unit tests home
DecoderState, FeatureExtractor, TranscriptAssembler, and EndpointMode
parse tests now live in state.rs. now_timestamp monotonicity lives next
to the helper in inference/mod.rs. Engine keeps speech_endpoint policy.

Constraint: no production behavior change
Confidence: high
Scope-risk: narrow
